refactor(samples): Use spawn helper from subprocess.ts
