+++ packages/create-project/src/helpers/install.ts
@@ -1,6 +1,7 @@
 // Modified from: https://github.com/vercel/next.js/blob/2425f4703c4c6164cecfdb6aa8f80046213f0cc6/packages/create-next-app/helpers/install.ts
 import chalk from 'chalk';
-import spawn from 'cross-spawn';
+
+import { spawn } from './subprocess';
 
 interface InstallArgs {
   /**
@@ -35,75 +36,64 @@ export function install(
    * Yarn-specific command-line flags.
    */
   const yarnFlags: string[] = [];
-  /**
-   * Return a Promise that resolves once the installation is finished.
-   */
-  return new Promise((resolve, reject) => {
-    let args: string[];
-    const command: string = useYarn ? 'yarnpkg' : 'npm';
 
-    if (dependencies && dependencies.length) {
-      /**
-       * If there are dependencies, run a variation of `{displayCommand} add`.
-       */
-      if (useYarn) {
-        /**
-         * Call `yarn add --exact (--offline)? (-D)? ...`.
-         */
-        args = ['add', '--exact'];
-        if (!isOnline) args.push('--offline');
-        args.push('--cwd', root);
-        if (devDependencies) args.push('--dev');
-        args.push(...dependencies);
-      } else {
-        /**
-         * Call `npm install [--save|--save-dev] ...`.
-         */
-        args = ['install', '--save-exact'];
-        args.push(devDependencies ? '--save-dev' : '--save');
-        args.push(...dependencies);
-      }
-    } else {
-      /**
-       * If there are no dependencies, run a variation of `{displayCommand}
-       * install`.
-       */
-      args = ['install'];
-      if (useYarn) {
-        if (!isOnline) {
-          console.log(chalk.yellow('You appear to be offline.'));
-          console.log(chalk.yellow('Falling back to the local Yarn cache.'));
-          console.log();
-          args.push('--offline');
-        }
-      } else {
-        if (!isOnline) {
-          console.log(chalk.yellow('You appear to be offline.'));
-          console.log();
-        }
-      }
-    }
+  let args: string[];
+  const command: string = useYarn ? 'yarnpkg' : 'npm';
+
+  if (dependencies && dependencies.length) {
     /**
-     * Add any package manager-specific flags.
+     * If there are dependencies, run a variation of `{displayCommand} add`.
      */
     if (useYarn) {
-      args.push(...yarnFlags);
+      /**
+       * Call `yarn add --exact (--offline)? (-D)? ...`.
+       */
+      args = ['add', '--exact'];
+      if (!isOnline) args.push('--offline');
+      args.push('--cwd', root);
+      if (devDependencies) args.push('--dev');
+      args.push(...dependencies);
     } else {
-      args.push(...npmFlags);
+      /**
+       * Call `npm install [--save|--save-dev] ...`.
+       */
+      args = ['install', '--save-exact'];
+      args.push(devDependencies ? '--save-dev' : '--save');
+      args.push(...dependencies);
     }
+  } else {
     /**
-     * Spawn the installation process.
+     * If there are no dependencies, run a variation of `{displayCommand}
+     * install`.
      */
-    const child = spawn(command, args, {
-      stdio: 'inherit',
-      env: { ...process.env, ADBLOCK: '1', DISABLE_OPENCOLLECTIVE: '1' },
-    });
-    child.on('close', (code) => {
-      if (code !== 0) {
-        reject({ command: `${command} ${args.join(' ')}` });
-        return;
+    args = ['install'];
+    if (useYarn) {
+      if (!isOnline) {
+        console.log(chalk.yellow('You appear to be offline.'));
+        console.log(chalk.yellow('Falling back to the local Yarn cache.'));
+        console.log();
+        args.push('--offline');
       }
-      resolve();
-    });
+    } else {
+      if (!isOnline) {
+        console.log(chalk.yellow('You appear to be offline.'));
+        console.log();
+      }
+    }
+  }
+  /**
+   * Add any package manager-specific flags.
+   */
+  if (useYarn) {
+    args.push(...yarnFlags);
+  } else {
+    args.push(...npmFlags);
+  }
+  /**
+   * Spawn the installation process.
+   */
+  return spawn(command, args, {
+    stdio: 'inherit',
+    env: { ...process.env, ADBLOCK: '1', DISABLE_OPENCOLLECTIVE: '1' },
   });
 }
